pyama/pyproject.toml
[project]
name = "pyama"
version = "0.1.0"
description = "Core processing library for PyAMA"
authors = [{ name = "PyAMA Team" }]
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "scipy",
    "nd2",
    "pandas<3",
    "pydantic>=2.0",
    "scikit-image",
    "tifffile",
    "pyyaml",
    "pylibczirw>=5.0.0",
    "zarr",
]

[tool.uv]
package = true

[[tool.uv.index]]
name = "scm"
url = "https://pypi.scm.io/simple/"

[tool.uv.sources]
pylibczirw = { index = "scm" }
